requirements.txt
cudnn==7.6.5
cupy==6.0.0
h5py==2.8.0
hdf5==1.8.20
keras-applications==1.0.8
keras-base==2.3.1
keras-gpu==2.3.1
keras-preprocessing==1.1.0
numpy==1.18.5
opencv==3.4.2
pillow==5.2.0
py-opencv==3.4.2
six==1.15.0
tensorflow-gpu==2.0.0
matplotlib==3.2.2
pandas==1.0.5
